+++ src/pages/Contact.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { InputForm } from "../components/InputForm";
+
+const Contact = () => {
+  return (
+    <div>
+      <InputForm />
+    </div>
+  );
+};
+
+export default Contact;
